pyproject.toml
[tool.poetry]
name = "langchain-qa_with_references"
version = "0.0.274"
description = "This is a temporary wproject while I wait for my langchain [pull-request](https://github.com/langchain-ai/langchain/pull/7278) to be validated."
authors = ["Philippe PRADOS"]
license = "MIT License"
readme = "README.md"
repository = "https://www.github.com/pprados/langchain-qa_with_references"
packages = [{include = "langchain_qa_with_references"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
langchain = "^0.0.274"
pydantic= "1.*"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
playwright = "^1.28.0"
setuptools = "^68.1.2"

[tool.poetry.group.test.dependencies]
# The only dependencies that should be added are
# dependencies used for running tests (e.g., pytest, freezegun, response).
# Any dependencies that do not meet that criteria will be removed.
pytest = "^7.3.0"
pytest-cov = "^4.0.0"
pytest-dotenv = "^0.5.2"
duckdb = "0.8.0"
#duckdb-engine = "^0.7.0"
#duckdb-engine = "0.8.0"
pytest-watcher = "^0.2.6"
freezegun = "^1.2.2"
responses = "^0.22.0"
pytest-asyncio = "^0.20.3"
lark = "^1.1.5"
pandas = "^2.0.0"
pytest-mock  = "^3.10.0"
pytest-socket = "^0.6.0"
#syrupy = "^4.0.2"
unstructured="^0.10.8"
twine="^4.0.0"
